requirements.txt
Flask==2.0.2
Flask-WTF==1.0.0
Flask-Executor==0.10.0
Flask-Cors==3.0.10
apispec==4.0.0
apispec-webframeworks==0.5.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.2
click==7.1.2
# normalize
nltk==3.6.5
polyglot==16.7.4
numpy==1.18.4
pandas==1.0.3
transliterate==1.10.2
Morfessor==2.0.6